+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 0f4e33f - Fix DAW/BAE/Unidad-*/Tarea-*
+| Author: fjrodafo - Sat, 15 Feb 2025 13:13:58 +0000
+| 
 * Commit: 9c12079 - Fix DAW/BAE/Unidad-*/Tarea-*
 | Author: fjrodafo - Sat, 15 Feb 2025 13:06:35 +0000
 |